+++ pmfp/entrypoint/grpc/build_/build_pb_js.py
@@ -6,7 +6,7 @@ from typing import List, Optional
 from pmfp.utils.run_command_utils import run
 
 
-def build_pb_js(files: List[str], includes: List[str], to: str, as_type: Optional[List[str]], cwd: Path,
+def build_pb_js(files: List[str], includes: List[str], to: str, cwd: Path,
                 ** kwargs: str) -> None:
     """编译grpc的protobuf定义文件为js语言模块.